I feel like I'm off to a pretty good start
I've been tweaking this setup for about a week, and I feel pretty good
about it. Right now I'm primarily using Emacs for Org Mode, but I'm
hoping to try to use it for some coding (aside from Lisp) soon.

For the longest time, I was using Emacs in the Terminal, which was
dumb. The GUI is so much nicer!

I wonder if I will regret the decision to check in my Tetris scores.