init-org-mobile: the files are always encrypted
Previously, encyption was conditional on the existence of the password
file; that kind of made sense when there was only one machine, but now
I am using Emacs on another machine, and the files ARE encrypted,
whether I have the password file or not. So, I changed the
configuration to reflect that, and also made the init/kill hooks
conditional, because it doesn't make sense to "automatically"
pull/push if it needs to prompt for a password.